course/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==3.2.19
django-cors-headers==4.0.0
django-six==1.0.5
djangorestframework==3.14.0
isoweek==1.3.3
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3